+++ wrangle/col/col_groupby_stats.py
@@ -1,18 +1,18 @@
 def col_groupby_stats(data, col, y):
-    
+
     """Takes in category column and returns
-    descriptive statistics based on an 
+    descriptive statistics based on an
     binary outcome feature.
-    
+
     data :
     col :
     y :
-    
+
     """
 
     import pandas as pd
     import wrangle as wr
-    
+
     data = data.copy(deep=True)
     data = data[[col, y]]
 
@@ -20,7 +20,7 @@ def col_groupby_stats(data, col, y):
     b = data.groupby(col).sum()
     d = data.groupby(col).mean()
     e = data.groupby(col).std()
-    
+
     a.columns = ['n']
     b.columns = ['sum']
     d.columns = ['mean']
@@ -32,5 +32,5 @@ def col_groupby_stats(data, col, y):
     out = wr.df_merge(d, out)
     out = wr.df_merge(e, out)
     out = out.round(2)
-    
+
     return out
