+++ wrangle/df/df_rename_col.py
@@ -1,14 +1,14 @@
 def df_rename_col(data, col, rename_to, destructive=False):
-    
+
     """Rename a single column
-    
+
     data : pandas DataFrame
         Pandas dataframe with the column to be renamed.
     col : str
         Column to be renamed
     rename_to : str
         New name for the column to be renamed
-    
+
     destructive : bool
         If set to True, will make changes directly to the dataframe which
         may be useful with very large dataframes instead of making a copy.
@@ -17,12 +17,11 @@ def df_rename_col(data, col, rename_to, destructive=False):
 
     if destructive is False:
         data = data.copy(deep=True)
-    
+
     cols = list(data.columns)
     loc = cols.index('index')
     cols.insert(loc, rename_to)
     cols.remove(col)
     data.columns = cols
-    
+
     return data
-
