+++ wrangle/dic/dic_corr_perc.py
@@ -1,8 +1,8 @@
 def dic_corr_perc(data, y):
-    
+
     """Performans a cut/bucket based correlation of values of all columns
     across multiple dataframes stored in a dictionary.
-    
+
     USE: dict_corr_perc(dfs, y)
 
     data : dict
@@ -18,31 +18,32 @@ def dic_corr_perc(data, y):
     final = pd.DataFrame()
 
     for label in list(data.keys()):
-	    for col in data[label].columns:
-	        if col != y:
+        for col in data[label].columns:
+            if col != y:
 
-	            # avoid destruction (yes this is hacky)
-	            buckets = wr.col_to_buckets(data[label], col)
-	            temp_df = pd.DataFrame({col: buckets, y: data[label][y]})
-	            
-	            # create a temp dataframe with correlation values
-	            temp_corrs = wr.col_corr_category(temp_df, col, y)
-	            temp_corrs = temp_corrs.reset_index()
-	            temp_corrs.columns = ['index', y] 
-	            
-	            # create a temp dataframe with sample counts
-	            temp_count = temp_df[col].value_counts().reset_index()
-	            temp_count.columns = ['index', 'samples']
-	            temp_count['index'] = temp_count['index'].astype('O')
-	            
-	            # combine correlation values and sample counts
-	            out = pd.merge(temp_corrs, temp_count, left_on='index', right_on='index')
+                # avoid destruction (yes this is hacky)
+                buckets = wr.col_to_buckets(data[label], col)
+                temp_df = pd.DataFrame({col: buckets, y: data[label][y]})
 
-	            # add handle columns
-	            out['metric'] = col
-	            out['metric_context'] = label
+                # create a temp dataframe with correlation values
+                temp_corrs = wr.col_corr_category(temp_df, col, y)
+                temp_corrs = temp_corrs.reset_index()
+                temp_corrs.columns = ['index', y]
 
-	            # append to the output dataframe 
-	            final = final.append(out)
+                # create a temp dataframe with sample counts
+                temp_count = temp_df[col].value_counts().reset_index()
+                temp_count.columns = ['index', 'samples']
+                temp_count['index'] = temp_count['index'].astype('O')
+
+                # combine correlation values and sample counts
+                out = pd.merge(temp_corrs, temp_count,
+                               left_on='index', right_on='index')
+
+                # add handle columns
+                out['metric'] = col
+                out['metric_context'] = label
+
+                # append to the output dataframe
+                final = final.append(out)
 
     return final
